auth-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
